+++ src/Sulu/Component/Websocket/MessageDispatcher/MessageDispatcher.php
@@ -45,13 +45,19 @@ class MessageDispatcher implements MessageDispatcherInterface
             throw new HandlerNotFoundException($name);
         }
 
-        $result = $this->handler[$name]->handle($conn, $message, $context);
+        $error = false;
+        try {
+            $message = $this->handler[$name]->handle($conn, $message, $context);
+        } catch (MessageHandlerException $ex) {
+            $message = $ex->getResponseMessage();
+            $error = true;
+        }
 
-        return $result !== null ?
-            array(
-                'handler' => $name,
-                'message' => $result,
-                'options' => $options
-            ) : null;
+        return array(
+            'handler' => $name,
+            'message' => $message,
+            'options' => $options,
+            'error' => $error
+        );
     }
 }